.bash_functions -> .shell_functions

+++ install.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 set -e
+DRY_RUN=${2:-"false"}
 
-BASE_DIR=$(cd `dirname "${BASH_SOURCE[0]}"` && pwd)
+BASE_DIR=$(cd $(dirname "${BASH_SOURCE[0]}") && pwd)
 
 source $BASE_DIR/.bash_colors
 
@@ -12,9 +13,14 @@ installVim(){
 }
 
 installBash(){
-	_install ".bashrc .bash_aliases .bash_functions .bash_colors" "bash"
+	_install ".bashrc .shell_aliases .shell_functions .bash_colors" "bash"
 	_install ".screenrc" "screen"
 	_install ".inputrc" "inputrc"
+	_install ".tmux.conf" "tmux"
+}
+installZsh(){
+	_install ".zshrc .shell_functions .shell_aliases" "zsh"
+	_install ".screenrc" "screen"
 	_install ".tmux.conf" "tmux"
 }
 installGit(){
@@ -25,7 +31,11 @@ installSlate(){
 }
 installKarabiner(){
 	echo -ne "\033[00;${COLOR_WHITE}mInstalling karabiner configuration file...\033[0m"
-	ln -sf "${BASE_DIR}/karabiner/private.xml" "${TARGET}/Library/Application Support/Karabiner/private.xml"
+	if [ "${DRY_RUN}" == "false" ]; then
+		ln -sf "${BASE_DIR}/karabiner/private.xml" "${TARGET}/Library/Application Support/Karabiner/private.xml"
+	else
+		echo Executing: ln -sf "${BASE_DIR}/karabiner/private.xml" "${TARGET}/Library/Application Support/Karabiner/private.xml"
+	fi
 	echo -e "[\033[01;${COLOR_GREEN}mDone\033[0m]"
 }
 installTerminalStuff(){
@@ -37,23 +47,28 @@ installIrssi(){
 _install(){
 	echo -ne "\033[00;${COLOR_WHITE}mInstalling $2 configuration...\033[0m"
 	for f in $1; do
-		ln -sf "${BASE_DIR}/${f}" "${TARGET}/${f}"
+		if [ "${DRY_RUN}" == "false" ]; then
+			ln -sf "${BASE_DIR}/${f}" "${TARGET}/${f}"
+		else
+			echo Executing: ln -sf "${BASE_DIR}/${f}" "${TARGET}/${f}"
+		fi
 	done
 	echo -e "[\033[01;${COLOR_GREEN}mDone\033[0m]"
 }
 case "$1" in
 	"dev-linux")
-		bash $0 "bash"
+		bash $0 "bash" ${DRY_RUN}
 		installGit
 		installVim
 	;;
 	"mac")
-		bash $0 "dev-linux"
-		installSlate
+		bash $0 "dev-linux" ${DRY_RUN}
+		installZsh
+#		installSlate
 #		installKarabiner
 	;;
 	"linux")
-		bash $0 "dev-linux"
+		bash $0 "dev-linux" ${DRY_RUN}
 		installIrssi
 	;;
 	"bash")
@@ -61,7 +76,7 @@ case "$1" in
 		installTerminalStuff
 	;;
 	*)
-		echo "$0 (dev-linux | mac | linux | bash)"
+		echo "$0 (dev-linux | mac | linux | bash) <?DRY_RUN:true/false>"
 		exit 1
 	;;
 esac
